Capitol3/Part2/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.2.2
zipp==3.11.0
